Removed tests for versions of the pg package under 6. Removed "dead" test code and rearranged "live" test code.
