--- test/probes/pg/basic.js
@@ -1,5 +1,0 @@
-'use strict'
-
-exports.run = function (ctx, done) {
-  ctx.pg.db.query('SELECT $1::int AS number', ['1'], done)
-}
